+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 # Change Log
 
-- added multiple files selection and destination directory selection to GUI
-- added support for PNG, JPG, and JPEG files
-- Using Zip to replace Word images instead of [python-docx](https://python-docx.readthedocs.io/en/latest/)
+- added remove using color replacements strategy
+- added flag to GUI to remove using color replacements strategy